feat: Updated routes and test to meet specs
