+++ transfer-connect/tests/test.rest
@@ -1,7 +1,7 @@
 ### Testing: route for bank server to send transfer information to TC
 ### Input: JSON file with transfer information
 ### Expected Output: JSON message of successful outcome
-POST http://localhost:3002/api/bank/send-transfer
+POST http://localhost:3002/api/bank/accrual-req
 Content-Type: application/json
 
 {
@@ -10,51 +10,58 @@ Content-Type: application/json
     "memberLastName": "Doe",
     "amount": 200,
     "referenceNumber": 2020080101,
-    "partnerCode": "DBSSG"
+    "partnerCode": "DBSSGXXXX"
 }
 
 ### Testing: route for bank server to get information about previously sent transfer from TC
-### Input: transfer id (id of collection in MongoDB) in the url
+### Input: JSON file containing transfer id (id of collection in MongoDB)
 ### Output: JSON file with transfer information
-GET http://localhost:3002/api/bank/get-transfer/62c16bd657cd7abc0d1712a7
+POST http://localhost:3002/api/bank/accrual-update
+Content-Type: application/json
+
+{
+    "accrual_refs": "62c16bd657cd7abc0d1712a7",
+    "partner_code": "DBSSG",
+    "api_key": "To Be Implemented"
+}
 
 
 ### Testing: route for bank server to get information about loyalty program
 ### Input: loyalty program id in the url
 ### Output: JSON file with loyalty program information
-GET http://localhost:3002/api/bank/get-loyalty-data/GOPOINTS
+GET http://localhost:3002/api/program/info/GOPOINTS
 
 ### Testing: route for bank server to validate loyalty points id, given program
 ### Input: program id, followed by loyalty points id in the url
 ### Output: JSON file true/false boolean value 
-GET http://localhost:3002/api/bank/get-LPID-validation/GOPOINTS/1234567890
+GET http://localhost:3002/api/program/validate/GOPOINTS/1234567890
 
 ### Testing: route for bank server to validate loyalty points id, given program
 ### Input: program id, followed by loyalty points id in the url
 ### Output: JSON file true/false boolean value 
-GET http://localhost:3002/api/bank/get-LPID-validation/INDOMILES/1234567890
+GET http://localhost:3002/api/program/validate/INDOMILES/1234567890
 
 ### Testing: route for bank server to validate loyalty points id, given program
 ### Input: program id, followed by loyalty points id in the url
 ### Output: JSON file true/false boolean value 
-GET http://localhost:3002/api/bank/get-LPID-validation/EMINENTGUEST/123456789012
+GET http://localhost:3002/api/program/validate/EMINENTGUEST/123456789012
 
 ### Testing: route for bank server to validate loyalty points id, given program
 ### Input: program id, followed by loyalty points id in the url
 ### Output: JSON file true/false boolean value 
-GET http://localhost:3002/api/bank/get-LPID-validation/QFLYER/1234567890
+GET http://localhost:3002/api/program/validate/QFLYER/1234567890
 
 ### Testing: route for bank server to validate loyalty points id, given program
 ### Input: program id, followed by loyalty points id in the url
 ### Output: JSON file true/false boolean value 
-GET http://localhost:3002/api/bank/get-LPID-validation/CONRADCLUB/123456789
+GET http://localhost:3002/api/program/validate/CONRADCLUB/123456789
 
 ### Testing: route for bank server to validate loyalty points id, given program
 ### Input: program id, followed by loyalty points id in the url
 ### Output: JSON file true/false boolean value 
-GET http://localhost:3002/api/bank/get-LPID-validation/MILLENIUMREWARDS/123456789L
+GET http://localhost:3002/api/program/validate/MILLENIUMREWARDS/123456789L
 
 ### Testing: route for bank server to validate loyalty points id, given program
 ### Input: program id, followed by loyalty points id in the url
 ### Output: JSON file true/false boolean value 
-GET http://localhost:3002/api/bank/get-LPID-validation/DOESNTEXIST/123456789
+GET http://localhost:3002/api/program/validate/DOESNTEXIST/123456789
